findById

+++ src/main/java/com/example/demo/repository/CustomerRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.model.Customer;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface CustomerRepository extends ReactiveCrudRepository<Customer, Integer> {
+public interface CustomerRepository extends ReactiveCrudRepository<Customer, Long> {
 
 }
